requirements/base.txt
Django==3.2.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-gis==0.17
django-cors-headers==3.7.0
dj-database-url==0.5.0
whitenoise==5.2.0
django-storages[azure]==1.11.1
sentry-sdk==1.3.1
requests==2.26.0
model-bakery==1.3.2
